+++ CHANGELOG.md
@@ -30,3 +30,6 @@ All notable changes to this project will be documented in this file.
 * RegexServiceImpl.getMatcher uses incorrect timeout [CTC-1629]
 * Web.download now creates all required folders [CTC-1532]
 * Allow custom character sets in Stream.write [CTC-1628]
+* Display a descriptive error when a concurrent modification occurs [CTC-1650]
+* File exists dialog does not overwrite file [CTC-1647]
+* Enable change of casing in robot name [CTC-1444]